+++ lofty/src/id3/v2/items/language_frame.rs
@@ -232,10 +232,10 @@ impl<'a> CommentFrame<'a> {
 	}
 }
 
-impl CommentFrame<'static> {
-	pub(crate) fn downgrade(&self) -> CommentFrame<'_> {
+impl CommentFrame<'_> {
+	pub(crate) fn borrow(&self) -> CommentFrame<'_> {
 		CommentFrame {
-			header: self.header.downgrade(),
+			header: self.header.borrow(),
 			encoding: self.encoding,
 			language: self.language,
 			description: Cow::Borrowed(&self.description),
@@ -365,10 +365,10 @@ impl<'a> UnsynchronizedTextFrame<'a> {
 	}
 }
 
-impl UnsynchronizedTextFrame<'static> {
-	pub(crate) fn downgrade(&self) -> UnsynchronizedTextFrame<'_> {
+impl UnsynchronizedTextFrame<'_> {
+	pub(crate) fn borrow(&self) -> UnsynchronizedTextFrame<'_> {
 		UnsynchronizedTextFrame {
-			header: self.header.downgrade(),
+			header: self.header.borrow(),
 			encoding: self.encoding,
 			language: self.language,
 			description: Cow::Borrowed(&self.description),